rag-chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
fastembed
qdrant_client
python-dotenv
chainlit==1.3.2